refactor: unify OSS update path through OSSUpdateStrategy
Move client-side OSS logic (version config download, upgrade check,
process launch) from LaunchOssAsync() into OSSUpdateStrategy. The
strategy now handles both client and upgrade OSS roles internally.

- OSSUpdateStrategy.ExecuteAsync(): detects role via
  GlobalConfigInfoOSS env var, dispatches to ExecuteClientAsync()
  or ExecuteUpgradeAsync() accordingly
- GeneralUpdateBootstrap: AppType.OSS now goes through
  LaunchWithStrategy(new OSSUpdateStrategy()) — consistent with
  Client and Upgrade paths
- Deleted LaunchOssAsync(), DownloadOssFile(), IsOssUpgrade() from
  GeneralUpdateBootstrap

Closes #409
